requirements.txt
# frameworks
django==2.1.7
djangorestframework==3.9.2
social-auth-app-django==3.1.0

# tools
ipython==7.4.0
